Refactor <annotation-overlay> and <annotation-list> to inherit from common class
